refactor Material enum. better now
